general cleanup of app and world map components
