+++ src/model/Cities.js
@@ -1,0 +1,8 @@
+const cities = ["london","paris","rome","oslo",
+"reykjavik","new-york","madrid","marrakech",
+"cairo","nairobi","istanbul" ,"dubai","cape-town",
+"los-angeles","mexico-city","bogota","rio-de-janeiro",
+"tehran","new-dehli","bangkok","shanghai","tokyo",
+"hong-kong","melbourne","wellington", "vancouver"]
+
+export default cities;
